+++ games/views.py
@@ -181,8 +181,9 @@ def list_sessions(
     context["dataset"] = dataset
     # cannot use dataset[0] here because that might be only partial QuerySet
     context["last"] = Session.objects.all().order_by("timestamp_start").last()
-    # charts are always oldest->newest
-    if dataset.count() >= 2:
+    # only if 2 or more non-manual sessions exist
+    if dataset.filter(timestamp_end__isnull=False).count() >= 2:
+        # charts are always oldest->newest
         context["chart"] = playtime_over_time_chart(dataset.order_by("timestamp_start"))
 
     return render(request, "list_sessions.html", context)