+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Improve form appearance
 * Add helper buttons next to datime fields
 * Change recent session view to current year instead of last 30 days
+* Fix bug when filtering only manual sessions (https://git.kucharczyk.xyz/lukas/timetracker/issues/51)
 
 ## 1.0.3 / 2023-02-20 17:16+01:00